change garnish.go ServeHTTP() function: change the connection between garnish and server from http to SCION
