+++ garnish/garnish.go
@@ -1,9 +1,14 @@
 package garnish
 
 import (
+	"context"
+	"fmt"
+	"github.com/netsec-ethz/scion-apps/pkg/pan"
+	"inet.af/netaddr"
 	"net/http"
 	"net/http/httputil"
 	"net/url"
+	"time"
 )
 
 const Xcache = "X-Cache"
@@ -30,8 +35,10 @@ func New(url url.URL) *garnish {
 	return &garnish{c: newCache(), proxy: reverseProxy}
 }
 
-func (g *garnish) ServeHTTP(rw http.ResponseWriter, r *http.Request) {
+func (g *garnish) ServeHTTP(rw http.ResponseWriter, r *http.Request, serverAddress string) {
 	// only GET requests should be cached
+	// send response back to the client
+	// do not need to change
 	if r.Method != http.MethodGet {
 		rw.Header().Set(Xcache, XcacheMiss)
 		g.proxy.ServeHTTP(rw, r)
@@ -40,24 +47,65 @@ func (g *garnish) ServeHTTP(rw http.ResponseWriter, r *http.Request) {
 
 	u := r.URL.String()
 	cached := g.c.get(u)
+
 	//if cached, return the cached data
+	// send response back to the client
+	// do not need to change
 	if cached != nil {
 		rw.Header().Set(Xcache, XcacheHit)
 		_, _ = rw.Write(cached)
 		return
 	}
 
-	proxyRW := &responseWriter{
-		proxied: rw,
+	// if not cached, should ask for the original server
+	// change from http connection to scion
+
+	// for a server, the proxy is like a client, so:
+	addr, err := pan.ResolveUDPAddr(serverAddress)
+	if err != nil {
+		fmt.Println("server address error")
+		return
 	}
-
-	proxyRW.Header().Set(Xcache, XcacheMiss)
-	g.proxy.ServeHTTP(proxyRW, r)
-
+	// garnish connect to the server
+	conn, err := pan.DialUDP(context.Background(), netaddr.IPPort{}, addr, nil, nil)
+	if err != nil {
+		fmt.Println("connect to server error")
+		return
+	}
+	defer conn.Close()
+	buffer := make([]byte, 16*1024)
+	if err = conn.SetReadDeadline(time.Now().Add(1 * time.Second)); err != nil {
+		fmt.Println("SetReadDeadline error")
+		return
+	}
+	n, err := conn.Read(buffer)
+	if err != nil {
+		fmt.Println("read from server error")
+		return
+	}
+	data := buffer[:n]
 	cc := rw.Header().Get(cacheControl)
 	toCache, duration := parseCacheControl(cc)
 	//check if it needs cache
 	if toCache {
-		g.c.store(u, proxyRW.body, duration)
+		g.c.store(u, data, duration)
 	}
+
+	//if not cached, should ask for the original server
+	//change to scion
+
+	//proxyRW := &responseWriter{
+	//	proxied: rw,
+	//}
+	//
+	//proxyRW.Header().Set(Xcache, XcacheMiss)
+	//g.proxy.ServeHTTP(proxyRW, r)
+	//
+	//cc := rw.Header().Get(cacheControl)
+	//toCache, duration := parseCacheControl(cc)
+	////check if it needs cache
+	//if toCache {
+	//	g.c.store(u, proxyRW.body, duration)
+	//}
+
 }
