+++ go.mod
@@ -2,4 +2,8 @@ module github.com/bkielbasa/garnish
 
 go 1.13
 
-require github.com/stretchr/testify v1.4.0
+require (
+	github.com/d4l3k/messagediff v1.2.1 // indirect
+	github.com/netsec-ethz/scion-apps v0.5.0
+	github.com/stretchr/testify v1.7.0
+)
